+++ core-shared-ui/src/main/java/com/codandotv/streamplayerapp/core_shared_ui/utils/Sharing.kt
@@ -1,5 +1,10 @@
 package com.codandotv.streamplayerapp.core_shared_ui.utils
 
+import android.content.Context
+import android.content.pm.PackageInfo
+import android.content.pm.PackageManager
+import android.os.Build
+
 object Sharing {
 
     const val SHARING_DATA_TYPE_TEXT = "text/plain"
@@ -9,3 +14,20 @@ object Sharing {
     const val OPTIONS_TITLE_MESSAGE = "Compartilhar usando"
     const val OPTIONS_NAME_MESSAGE = "Link Download"
 }
+
+fun isPackageInstalled(packageName: String, context: Context): Boolean {
+    val pm = context.packageManager
+    return try {
+        pm.getPackageInfoCompat(packageName)
+        true
+    } catch (e: PackageManager.NameNotFoundException) {
+        false
+    }
+}
+
+fun PackageManager.getPackageInfoCompat(packageName: String, flags: Int = 0): PackageInfo =
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+        getPackageInfo(packageName, PackageManager.PackageInfoFlags.of(flags.toLong()))
+    } else {
+        @Suppress("DEPRECATION") getPackageInfo(packageName, flags)
+    }
